Move policy decision types to permission-common
