+++ plugins/permission-common/src/types/index.ts
@@ -22,6 +22,9 @@ export type {
   AuthorizeResponse,
   IdentifiedPermissionMessage,
   PermissionMessageBatch,
+  ConditionalPolicyDecision,
+  DefinitivePolicyDecision,
+  PolicyDecision,
   PermissionCondition,
   PermissionCriteria,
   AllOfCriteria,
